# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ThePower LMS API
  description: API multi-tenant para plataforma LMS B2B con FUNDAE compliance
  version: 1.0.0
  contact:
    name: ThePower Education
servers:
  - url: https://api.thepower.education/v1
    description: Production
  - url: https://api.staging.thepower.education/v1
    description: Staging
security:
  - BearerAuth: []
  - OAuth2: [read, write]
...
paths:
...
  /courses:
    post:
      tags: [Courses]
      summary: Crear curso
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        '201':
          description: Curso creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
  /courses/{courseId}:
    put:
      tags: [Courses]
      summary: Actualizar curso
      operationId: updateCourse
      parameters:
        - $ref: '#/components/parameters/courseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdate'
      responses:
        '200':
          description: Curso actualizado
...
  /courses/{courseId}/modules:
    post:
      tags: [Courses]
      summary: Agregar módulo al curso
      operationId: addModule
      parameters:
        - $ref: '#/components/parameters/courseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleCreate'
      responses:
        '201':
          description: Módulo agregado

  # === ENROLLMENTS ===
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            read: Read access
            write: Write access
  parameters:
...
    courseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        hours:
          type: number
        level:
          type: string
        familyProfessional:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        fundaeCompatible:
          type: boolean
        version:
          type: integer
    CourseCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        hours:
          type: number
        level:
          type: string
        familyProfessional:
          type: string
        fundaeCompatible:
          type: boolean
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        hours:
          type: number
        level:
          type: string
    ModuleCreate:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [video, pdf, text, quiz, scorm]
        contentUrl:
          type: string
        durationMinutes:
          type: integer
        orderIndex:
          type: integer
        isMandatory:
          type: boolean
...